delete version.go
